Remove VTE 'clean' flag
This flag blocks "Set Path From Document" and "Execute" commands
if terminal was once in touch. It is not possible to make it work
well in all cases, but sometimes it prevents a normal workflow.
Anyway it doesn't cover cases of ssh, interactive python/cat,
Ctrl+D and so on.
Fixes #542